+++ test/bubble.js
@@ -3,7 +3,7 @@
 gadgetui.objects.Constructor( gadgetui.display.Bubble, [document.getElementsByTagName( "select" )[0],
 	"Select Food",
 	{
-		position : "middle right",
+		position : "middle center",
 		arrowSize: 0,
 		class: "bubbleClass",
 		closable : true,